% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate}
\alias{generate}
\title{Generate Mother-Child Linkages}
\usage{
generate(connectionDetails, cdmDatabaseSchema, resultsDatabaseSchema,
  pppDatabaseSchema = cdmDatabaseSchema, pppTableName = "payer_plan_period",
  motherRelationshipId = 40478925, childRelationshipId = 40485452,
  ffdf = NULL)
}
\arguments{
\item{connectionDetails}{An R object of type ConnectionDetails (details for the function that contains server info, 
database type, optionally username/password, port)}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM (including Vocabulary). 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that holds the pregnancy_episodes table.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{pppDatabaseSchema}{Fully qualified name of database schema that holds the payer_plan_period table with family_source_value populated. 
Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, 
so for example, on SQL Server, 'cdm_scratch.dbo'.}

\item{pppTableName}{(OPTIONAL) The name of the table to land the custom payer_plan_period table. By default, this is "payer_plan_period."}

\item{motherRelationshipId}{(OPTIONAL) The concept Id to relate a mother to a child. By default, it is 40478925.}

\item{childRelationshipId}{(OPTIONAL) The concept Id to relate a child to a parent. By default, it is 40485452.}

\item{ffdf}{(OPTIONAL) An ffdf object with family ids -- only use this if your CDM payer_plan_period table doesn't
have family ids}
}
\description{
\code{generate} creates descriptive statistics summary for an entire OMOP CDM instance.
}
\details{
\code{generate} creates descriptive statistics summary for an entire OMOP CDM instance.
}
