% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getFamilyIds}
\alias{getFamilyIds}
\title{Get family IDs from native data set}
\usage{
getFamilyIds(nativeConnectionDetails, nativeDatabaseSchema, nativeTable,
  nativePersonId, nativeFamilyId, nativeStartDate, nativeEndDate)
}
\arguments{
\item{nativeConnectionDetails}{An R object of type ConnectionDetails (details for the function that contains server info, database type, optionally username/password, port)}

\item{nativeDatabaseSchema}{Fully qualified name of database schema that contains the native data set.}

\item{nativeTable}{The name of the table that holds the family Id in the native data set.}

\item{nativePersonId}{The name of the person_id field in the native data set.}

\item{nativeFamilyId}{The name of the family_id field in the native data set.}

\item{nativeStartDate}{The name of the start date field for the observation period / payer plan period}

\item{nativeEndDate}{The name of the end date field for the observation period / payer plan period}
}
\value{
ffdf object with the person - family Id linkage
}
\description{
\code{getFamilyIds} obtains the family IDs from the native data source and writes to ff object.
}
\details{
\code{getFamilyIds} obtains the family IDs from the native data source, which is useful if family IDs are time-varying and cannot
                    be stored in the payer_plan_period.family_source_value field.
}
